requirements.txt
rank_bm25
rank-llm==0.12.8
chromadb
langchain
langchain-core
langchain-community
langchain-openai
langchain_cohere
docling
cohere
pypdf
python-dotenv
pyyaml
chroma-hnswlib
regex